requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
numpy
pillow
python-dotenv
opencv-python-headless
insightface
python-multipart
langchain
onnxruntime